refactor: extract `useSlideInfo`
